Removing lots of useless (int) casts

+++ Algorithms/lis_2d.cpp
@@ -53,12 +53,12 @@ void build_x(int cur, int l, int r){
 
 	if (l == r){ // Creating leaf and building segment tree on y coordinate.
 		// Inserting every unique sorted y coordinate that has vx[l] as x coordinate.
-		for (i = 0; i < (int)mx[l].size(); i++){
+		for (i = 0; i < mx[l].size(); i++){
 			seg_x[cur].push_back(mx[l][i]);
 		}
 
 		// Allocating memory for the segment tree stored in this leaf.
-		seg_y[cur].resize(4 * (int)seg_x[cur].size() + 1);
+		seg_y[cur].resize(4 * seg_x[cur].size() + 1);
 		build_y(seg_y[cur], seg_x[cur], 1, 0, seg_x[cur].size() - 1);
 		return;
 	}
@@ -69,7 +69,7 @@ void build_x(int cur, int l, int r){
 
 	// Merging and building segment tree on y coordinates.
 	seg_x[cur] = merge_x(seg_x[LEFT(cur)], seg_x[RIGHT(cur)]);
-	seg_y[cur].resize(4 * (int)seg_x[cur].size() + 1);
+	seg_y[cur].resize(4 * seg_x[cur].size() + 1);
 	build_y(seg_y[cur], seg_x[cur], 1, 0, seg_x[cur].size() - 1);
 }
 
@@ -164,7 +164,7 @@ void init(){
 	int pos;
 
 	// Retrieving x coordinates.
-	for (int i = 0; i < (int)p.size(); i++){
+	for (int i = 0; i < p.size(); i++){
 		vx.push_back(p[i].x);
 	}
 
@@ -175,19 +175,19 @@ void init(){
 	// Retrieving y coordinates for every x coordinate.
 	mx.resize(vx.size());
 
-	for (int i = 0; i < (int)p.size(); i++){
+	for (int i = 0; i < p.size(); i++){
 		pos = lower_bound(vx.begin(), vx.end(), p[i].x) - vx.begin();
 		mx[pos].push_back(p[i].y);
 	}
 
 	// Making y coordinates unique for every x coordinate.
-	for (int i = 0; i < (int)mx.size(); i++){
+	for (int i = 0; i < mx.size(); i++){
 		sort(mx[i].begin(), mx[i].end());
 		mx[i].resize(unique(mx[i].begin(), mx[i].end()) - mx[i].begin());
 	}
 
 	// Building.
-	build_x(1, 0, (int)vx.size() - 1);
+	build_x(1, 0, vx.size() - 1);
 }
 
 /* O(N * Log^2(N)). */
@@ -196,9 +196,9 @@ int lis_2d(){
 	dp.resize(p.size());
 
 	// Filling dp.
-	for (int i = 0; i < (int)p.size(); i++){
-		dp[i] = 1 + query_x(1, 0, (int)vx.size() - 1, 1, p[i].x, 1, p[i].y);
-		update_x(1, 0, (int)vx.size() - 1, p[i].x, p[i].y, dp[i]);
+	for (int i = 0; i < p.size(); i++){
+		dp[i] = 1 + query_x(1, 0, vx.size() - 1, 1, p[i].x, 1, p[i].y);
+		update_x(1, 0, vx.size() - 1, p[i].x, p[i].y, dp[i]);
 	}
 
 	// Returning LIS.
