+++ Geometry/shoelace.cpp
@@ -15,7 +15,7 @@ double shoelace(const vector<Point> &p){
 
 	ans = 0;
 
-	for (int i = 0; i < (int)p.size() - 1; i++){
+	for (int i = 0; i < p.size() - 1; i++){
 		ans += p[i].x * p[i + 1].y - p[i + 1].x * p[i].y;
 	}
