+++ Graph Theory/cycles.cpp
@@ -14,7 +14,7 @@ void dfs(int u){
 	in_stack[u] = true;
 	seen[u] = true;
 
-	for (int i = 0; i < (int)g[u].size(); i++){
+	for (int i = 0; i < g[u].size(); i++){
 		int v = g[u][i];
 
 		if (!seen[v]){ // Not seen yet.
