+++ Graph Theory/lca_euler_tour.cpp
@@ -71,7 +71,7 @@ void dfs(int u, int d){
 	euler[cur] = u;
 	depth[cur] = d;
 
-	for (int i = 0; i < (int)g[u].size(); i++){
+	for (int i = 0; i < g[u].size(); i++){
 		int v = g[u][i];
 
 		if (!seen[v]){
