+++ Number Theory/power_tower_fermat_euler.cpp
@@ -133,14 +133,14 @@ void chinese_remainder_theorem(vector<long long> a, vector<long long> m, long lo
 	long long a2, m2, x1, y1, gcd, lcm;
 
 	// Making 0 <= ai < mi.
-	for (int i = 0; i < (int)a.size(); i++){
+	for (int i = 0; i < a.size(); i++){
 		a[i] = ((a[i] % m[i]) + m[i]) % m[i];
 	}
 
 	a1 = a[0];
 	m1 = m[0];
 
-	for (int i = 1; i < (int)a.size(); i++){
+	for (int i = 1; i < a.size(); i++){
 		a2 = a[i];
 		m2 = m[i];
